Try and make google log in work.
The Google log in process is a nightmarish mix of javascript, redirects,
cookies and SSL. This fix adds cache-control parsing to stop the caching
of some URLs, which keeps content fresh. Non-cached files are cached then
removed immediately, which is easier than not caching at all.

In theory the log works without javascript, but it seems to get
into an infinite loop. A maximum number of redirects is added to
prevent infinite recursion. This is set to 7 by default, but can be
altered in the config.ini file.

git-svn-id: https://quirkysoft.googlecode.com/svn/trunk@543 89bda9af-7d2e-0410-a434-4dd01d8ab487