Fix caching for internal link pages.
When a page has an internal link, strip off the #link part when calculating
the filename CRC. Then on loading a page with an internal link, get
the base page name and load this. Mark the internal link page as visited
so that the clicked link colour can be set.

This means internal link pages get a cache hit and are not reloaded
from the web site.


git-svn-id: https://quirkysoft.googlecode.com/svn/trunk@648 89bda9af-7d2e-0410-a434-4dd01d8ab487