Improve downloading to only connect once.
Some sites . Google Code, I am looking at you - don't respond to a HEAD
request, but do respond to a GET. This plays havoc with my previous
approach.

Change back to an approach similar to before, but now when the
HeaderParser reaches the end of the header section (\r\n\r\n), it notifies
anyone interested that the headers are ready - in this case the Document.
The Document then notifies *its* listeners that it now has the headers.

Once the View knows the headers it can act, if it is a non-viewable type,
the save-as screen is shown. Then if the user chooses a file name, the file
will be saved here. The Controller remembers the file name chosen and
does the actual "save" by copying from the cache to the chosen file name.
This save is performed either when the file finishes, or when the user
makes up their mind. Whichever takes longest.

Finally, always overwrite files when saving. A bit dangerous... need a
better way!


git-svn-id: https://quirkysoft.googlecode.com/svn/trunk@670 89bda9af-7d2e-0410-a434-4dd01d8ab487