Starting work on auto-updater.
The auto updater is going to be semi-automatic initially, it will require
the user to push a button. This will improve "later". This commit adds
the button to the config page and adds a dummy update URL for local tests.

Changes:
Config now reads the template file to get new values.
Controller checks the posted action to see if the user is changing config
 or checking updates.
URI, UnicodeString move the tokenize function from URI, fix stupid warning.
View add method to disable save-as for auto download.


git-svn-id: https://quirkysoft.googlecode.com/svn/trunk@726 89bda9af-7d2e-0410-a434-4dd01d8ab487