Add a mechanism for allowing cookies per site.
Fill in the View::addCookie() method so that the user can choose whether to
add the current site or domain to the allowed cookie list. A new handler
class, "CookieHandler", is added that has a title a bit of text and a
few buttons. It also has 2 radio buttons to allow the user to choose
site or domain.

Finally, change the std::map for a std::set in the CookieJar class.


git-svn-id: https://quirkysoft.googlecode.com/svn/trunk@800 89bda9af-7d2e-0410-a434-4dd01d8ab487