Switch of wifi connection by clicking the icon.
Sometimes I start reading a long blog post and I'm thinking of all the
juice the wifi stuff is sucking out the DS. With this change, clicking
the little antenna icon disables the Wifi hardware to save batteries.

Remove the 0xDEADFEED code path since it doesn't actually work. If the
Wifi is deinit'd then it never reconnects. Fix the PC version to allow
the disabled icon to show. Fix space errors in the arm* versions of the
Wifi classes.