Add an installation option.
The PREFIX/--prefix idea is nice, but requires reconfiguring the project.
Add a --install-to option that installs the .nds file and data directory
to the specified install location without needing a new configure.

Ensure that only the things we want to install get installed, using
the undocumented inst_var property.